ARM: EXYNOS: Fix argument of get_int_volt() bug
Parameter type of get_int_volt fuction is table index.
But caller on exynos_busfreq_reboot_event function uses freqency type.

+++ arch/arm/mach-exynos/busfreq_opp_4x12.c
@@ -581,7 +581,7 @@ int exynos4x12_find_busfreq_by_volt(unsigned int req_volt, unsigned int *freq)
 }
 EXPORT_SYMBOL_GPL(exynos4x12_find_busfreq_by_volt);
 
-unsigned int exynos4x12_get_int_volt(unsigned long index)
+unsigned int exynos4x12_get_int_volt(unsigned int index)
 {
 	return exynos4_int_volt[asv_group_index][index];
 }
